[CLEANUP] Moved Header, Main to Components [CLEANUP] Move App to Pages
