+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './stylesheets/Master.css';
 import registerServiceWorker from './registerServiceWorker';
 import { BrowserRouter } from 'react-router-dom'
-import App from "./components/App/App.jsx"
+import App from "./pages/App/App.jsx"
 
 // Renders App in the Router
 ReactDOM.render((
